+++ src/mutex.h
@@ -19,7 +19,7 @@ extern "C" {
 #endif
 
 os_status_t osi_mutex_create(os_mutex_t *mutex, os_mutex_definition_t *def);
-os_status_t osi_mutex_acquire(os_mutex_t *mutex, uint16_t to);
+os_status_t osi_mutex_acquire(os_mutex_t *mutex, uint32_t to);
 os_status_t osi_mutex_release(os_mutex_t *mutex);
 
 #if defined(__cplusplus)
